health mana and exp ui bars are in place

+++ Assets/Scripts/Scriptable Objects/ProjectileData.cs
@@ -6,6 +6,6 @@ public class ProjectileData : ScriptableObject
 {
     [field: SerializeField] public string Name { get; private set; }
     [field: SerializeField] public float Speed { get; private set; }
-    [field: SerializeField] public float Damage { get; private set; }
+    [field: SerializeField] public int Damage { get; private set; }
     [field: SerializeField] public float Lifespan { get; private set; }
 }
